backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.0
newsapi-python==0.2.7
google-genai==1.7.0
python-multipart==0.0.6
starlette==0.27.0
typing-extensions==4.8.0
aiohttp==3.8.6
langchain==0.0.350
openai==1.3.0
pypdf==3.17.1
beautifulsoup4==4.12.2
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 